Prevent entities from appearing in notifications (Closes #7)
Twitter changes < and > characters to their HTML entity versions.
This displays all &lt; and &gt; substrings in a tweet as < and >,
respectively. This means said substrings will never appear in
a notification, even if it's what the tweet actually was meant to
contain, but that is a result of Twitter's deliver of tweet text.